Handle nodes without a parentNode
**why**

React rendering during testing uses a detached DOM node. This was
causing issues with stickyfill because in 2 places (lines 314 and 549)
the DOM is traversed upwards until the document is reached. In a
detached DOM node, the document is never reached.

**what**

In these instances I manually jump to the body to continue traversal.